Synchronize google tasks and gtd gdocs (#72)
* Implement basic functionaliy of sync to google tasks.

There is one additional menu test tasks. After clicking it, a new task
will be added to the google tasks.

Note that there is still error when updating tasks

* Implement basic functionaliy of sync to google tasks.

There is one additional menu test tasks. After clicking it, a new task
will be added to the google tasks.

Note that there is still error when updating tasks

* When click menus gtasks will be updated automatically

* Update compiled_script.js in codebase

* Disable gtask sync if Google Tasks is not enabled

* Update Readme for sync to google tasks